Separated WebSearchTab and ExportTab
