+++ src/main/java/org/jabref/gui/preferences/export/ExportTab.fxml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.VBox?>
+<?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
+<fx:root spacing="10.0" type="VBox"
+         xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
+         fx:controller="org.jabref.gui.preferences.export.ExportTab">
+    <Label styleClass="titleHeader" text="%Export"/>
+
+    <Label styleClass="sectionHeader" text="%Export sort order"/>
+    <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
+</fx:root>
