+++ src/main/java/org/jabref/gui/preferences/export/ExportTab.java
@@ -1,0 +1,37 @@
+package org.jabref.gui.preferences.export;
+
+import javafx.fxml.FXML;
+
+import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel;
+import org.jabref.gui.preferences.AbstractPreferenceTabView;
+import org.jabref.gui.preferences.PreferencesTab;
+import org.jabref.logic.l10n.Localization;
+
+import com.airhacks.afterburner.views.ViewLoader;
+
+public class ExportTab extends AbstractPreferenceTabView<ExportTabViewModel> implements PreferencesTab {
+
+    @FXML private SaveOrderConfigPanel exportOrderPanel;
+
+    public ExportTab() {
+        ViewLoader.view(this)
+                  .root(this)
+                  .load();
+    }
+
+    @Override
+    public String getTabName() {
+        return Localization.lang("Export");
+    }
+
+    public void initialize() {
+        this.viewModel = new ExportTabViewModel(preferencesService);
+
+        exportOrderPanel.saveInOriginalProperty().bindBidirectional(viewModel.saveInOriginalProperty());
+        exportOrderPanel.saveInTableOrderProperty().bindBidirectional(viewModel.saveInTableOrderProperty());
+        exportOrderPanel.saveInSpecifiedOrderProperty().bindBidirectional(viewModel.saveInSpecifiedOrderProperty());
+        exportOrderPanel.sortableFieldsProperty().bind(viewModel.sortableFieldsProperty());
+        exportOrderPanel.sortCriteriaProperty().bindBidirectional(viewModel.sortCriteriaProperty());
+        exportOrderPanel.setCriteriaLimit(3);
+    }
+}
