+++ src/main/java/org/jabref/gui/preferences/websearch/WebSearchTab.fxml
@@ -8,7 +8,6 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.websearch.WebSearchTab">
@@ -24,9 +23,6 @@
         <CheckBox fx:id="useCustomDOI" text="%Use custom DOI base URI for article access"/>
         <TextField fx:id="useCustomDOIName" HBox.hgrow="ALWAYS"/>
     </HBox>
-
-    <Label styleClass="sectionHeader" text="%Export sort order"/>
-    <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
 
     <Label styleClass="sectionHeader" text="%Remote services"/>
     <CheckBox fx:id="grobidEnabled" text="%Allow sending PDF files and raw citation strings to a JabRef online service (Grobid) to determine Metadata. This produces better results."/>
